requirements.txt
customtkinter
Pillow
opencv-python
numpy
pytesseract
pdf2image
google-genai
python-dotenv
pyinstaller
scikit-image
img2pdf
